+++ src/org/zowe/pipelines/nodejs/exceptions/BuildStageException.groovy
@@ -1,0 +1,7 @@
+package org.zowe.pipelines.nodejs.exceptions
+
+class BuildStageException extends StageException {
+    BuildStageException(String message, String stageName) {
+        super(message, stageName)
+    }
+}
