+++ src/org/zowe/pipelines/nodejs/exceptions/NodejsRunnerException.groovy
@@ -1,0 +1,7 @@
+package org.zowe.pipelines.nodejs.exceptions
+
+class NodeJSRunnerException extends Exception {
+    NodeJSRunnerException(String message) {
+        super(message)
+    }
+}
