+++ src/org/zowe/pipelines/nodejs/exceptions/TestStageException.groovy
@@ -1,0 +1,7 @@
+package org.zowe.pipelines.nodejs.exceptions
+
+class TestStageException extends StageException {
+    TestStageException(String message, String stageName) {
+        super(message, stageName)
+    }
+}
